requirements.txt
autodoc-pydantic==1.5.0
celery==4.1.0
channels==3.0.3
channels-redis==3.2.0
django-anymail==8.4
django-sms[twilio,messagebird]==0.5.0
docutils<=0.16
django-jsonfield-backport==1.0.4
django-rq==2.4.0
m2r2==0.2.7
pre-commit==2.15.0
pusher==3.0.0
pydantic==1.8.2
pytz==2017.2
requests==2.25.1
slack_sdk==3.11.2
Sphinx==3.5.1
sphinx-rtd-theme==0.5.1
tweepy==4.1.0
